대화창 ui 변경, 스카이박스 구현, 깨진 하트 적용

+++ Assets/Scripts/HwangSuHo/Boss/White/WhiteMage.cs
@@ -38,7 +38,7 @@ public class WhiteMage : Entity
     }
     private void RenderHealth()
     {
-        BossHPGraphRenderer.Instance.Render(HP / MaxHP);
+        BossHPGraphRenderer.Instance.Render(HP / (float)MaxHP);
     }
 
     //50% 남았을때 인지 아닌지 확인
